perf: utilizar double em varivavel acumuladora

+++ src/cross_correlation.c
@@ -7,44 +7,44 @@ void calcular_correlacao_cruzada(float x[], float y[], float resultado[]) {
     int amostras_sobrepostas;
     int inicio_x;
     int inicio_y;
-    float soma;
+    double soma;
 
     for (i = 0; i < SIGNAL_SIZE - 1; i++) {
         deslocamento = (SIGNAL_SIZE - 1) - i;
         amostras_sobrepostas = i + 1;
         inicio_x = 0;
         inicio_y = deslocamento;
-        soma = 0.0f;
+        soma = 0.0;
 
         for (j = 0; j < amostras_sobrepostas; j++) {
-            soma += x[inicio_x + j] * y[inicio_y + j];
+            soma += (double) x[inicio_x + j] * (double) y[inicio_y + j];
         }
 
-        resultado[i] = soma;
+        resultado[i] = (float) soma;
     }
 
     amostras_sobrepostas = SIGNAL_SIZE;
     inicio_x = 0;
     inicio_y = 0;
-    soma = 0.0f;
+    soma = 0.0;
 
     for (j = 0; j < amostras_sobrepostas; j++) {
-        soma += x[inicio_x + j] * y[inicio_y + j];
+        soma += (double) x[inicio_x + j] * (double) y[inicio_y + j];
     }
 
-    resultado[SIGNAL_SIZE - 1] = soma;
+    resultado[SIGNAL_SIZE - 1] = (float) soma;
 
     for (i = SIGNAL_SIZE; i < CORRELATION_SIZE; i++) {
         deslocamento = i - (SIGNAL_SIZE - 1);
         amostras_sobrepostas = SIGNAL_SIZE - deslocamento;
         inicio_x = deslocamento;
         inicio_y = 0;
-        soma = 0.0f;
+        soma = 0.0;
 
         for (j = 0; j < amostras_sobrepostas; j++) {
-            soma += x[inicio_x + j] * y[inicio_y + j];
+            soma += (double) x[inicio_x + j] * (double) y[inicio_y + j];
         }
 
-        resultado[i] = soma;
+        resultado[i] = (float) soma;
     }
 }
